refactor: 💡 service layer folder reorganize
