+++ src/app/(not-logged-in)/activation-request/page.tsx
@@ -1,5 +1,4 @@
-import { NotLoggedIn } from '@/components';
-
 export default function ActivationRequestPage() {
-  return <NotLoggedIn.ActivationRequest />;
+  // return <NotLoggedIn.ActivationRequest />;
+  return <h1>Activation Request</h1>;
 }
