--- src/components/not-logged-in/activation-request/index.ts
@@ -1,1 +1,0 @@
-export { ActivationRequestViewModel as ActivationRequest } from './activation-request.viewmodel';
