+++ src/components/not-logged-in/index.ts
@@ -1,4 +1,3 @@
-export { ActivationRequest } from './activation-request';
 export { Container } from './container';
 export { FooterWaves } from './footer-waves';
 export { HeaderLogo } from './header-logo';
