+++ src/components/not-logged-in/signup/form/form.model.ts
@@ -4,8 +4,8 @@ import { useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
 
-import { SignupResponse } from '@/service/signup/signup.types';
-import { PersonType, Register } from '@/types/register';
+import { PersonType, Register } from '@/service/domain/register';
+import { SignupResponse } from '@/service/types';
 import { SIGNUP_FLOW } from '../signup.constants';
 import { PERSON_TYPES } from './form.constants';
 import { SIGNUP_FORM_SCHEMA } from './form.schema';
