+++ src/service/constants.ts
@@ -1,4 +1,4 @@
-import { API_COMMON_ERRORS } from '.';
+import { API_COMMON_ERRORS } from './types';
 
 export const API_COMMON_RESPONSE_ERROR: Record<API_COMMON_ERRORS, string> = {
   API_SERVER_ERROR:
