+++ src/service/index.ts
@@ -1,0 +1,7 @@
+import { activate } from './action/activate.service';
+import { signup } from './action/signup.service';
+
+export const CertFlowServices = {
+  signup,
+  activate
+};
