--- src/service/signup/index.ts
@@ -1,7 +1,0 @@
-import { activate } from './signup-activation.service';
-import { signup } from './signup.service';
-
-export const SignupService = {
-  activate,
-  signup
-};
